Bug fixed

+++ main.py
@@ -4,6 +4,7 @@ from read_lib import read_lib
 import ngender
 from ocr import ocr
 import difflib
+import threading
 
 CACHE_DIR = "./cache/"
 
@@ -63,7 +64,7 @@ while True:
             if in_dialogue():
                 print("In dialogue...")
                 person, text = get_dialogue()
-                if string_similar(text, last_text) < 0.9:
+                if string_similar(text, last_text) < 0.9 and text.endswith(('.', '!', '?', '。', '！', '？')):
                     os.system(f"edge-tts --rate=+10% --text '{text}' --voice {person_to_voice(person)} --write-media {CACHE_DIR}voice.wav")
                     os.system("mpv " + CACHE_DIR + "voice.wav")
                     os.remove(CACHE_DIR + "voice.wav")
@@ -71,6 +72,6 @@ while True:
     except KeyboardInterrupt:
         for file_name in os.listdir(CACHE_DIR):
             os.remove(CACHE_DIR + file_name)
-        break
+        exit()
     except:
         pass
